modules/base/scripts/requirements.txt
openai-whisper==20240107
numpy>=1.23.5
torch>=2.0.0
torchaudio>=2.0.0
ffmpeg-python>=0.2.0 